Challenge5 - Alaska Department of Snow Online Agent/requirements.txt
streamlit
python-dotenv
google-cloud-aiplatform
pandas
pytest
google-cloud-bigquery
google-cloud-aiplatform[evaluation]
pyarrow
db-dtypes